[IMP] Finished dispatching tests, remaining ones now in tests_rest_miscellaneous
